Enable ports 5900-5999 for HTML5 console
Removed the `firewall` command in favor of opening
ports in the post section.

The kickstart `firewall` command was adding the ports
to the default zone which is "public" at the time the
main section runs.

We later create the "manageiq" zone and set that to the
default causing the ports to still be closed in the active
zone.

https://bugzilla.redhat.com/show_bug.cgi?id=1278432