Backport Vagrant support from master.
This only brings back the changes to the kickstart files. The other
changes are too invasive to backport. So you need to build this with the
latest manageiq-appliance-build, as such:

  ruby vmbuild.rb -r darga-3 -t release

You need to be careful not to include the --local switch here, so that
the correct Darga kickstarts are used with the new automation.

(Cherry picked from commit ca57df70bd4dbb1a50dfa0eeaca07737727d093a)