Small build performance tweaks.
Using text mode reduces the build by about 2 minutes.

Also add the timestamp to PS4 so that we can see which postinstall
steps take the most time.

Increase the VM size to 2GB. We'll be changing the postinstall in the
subsequent patches to use a few tmpfs file systems. So make space for
that.