Don't include temporary and cache files in the image.
Save about 315 MiB of space (uncompressed) in the image by not including
temporary and cache files. These files are created by the postinstall
script. Note that we can't just remove the files because that would not
deallocate the storage. Instead, we mount a tmpfs on /tmp and /var/cache
as the first step int he postinstall. That way these files do not end up
in the image.