Shallow checkout of sources
When building from a tag, we do a shallow clone (git clone --depth 1). This
saves in excess of 100 MiB for the manageiq repository.

We don't do shallow clones when building from a branch. The reason for this is
that the result would be somewhat fragile. The build scripts decide which
commit sha to build before the install (so that the sha can be referenced in
the file name). During the post-install we're about 30 minutes later, and the
repository could have gotten additional commits. Checking out with --depth 1
will only allow us to git reset --hard to the latest commit.

A partial solution would be to clone with --depth 10 or something. While in
practise this would probably work, I don't like it. So therefore this patch
only does the shallow clone for tags. All release builds are built from tags,
and this is where it's most important to save space.