Add support for building for libvirt.
Theoretically libvirt supports multiple hypervisors, so it would be more
correct to call the target "libvirt-qemu". However I really really doubt
that we will ever support anything else. So take the same approach as
with the openstack target and just call it "libvirt" (openstack is
called "openstack" and not "openstack-kvm"). In the unlikely case there
will be a need to support another libvirt backend then that target can
be called "libvirt-foo".