vmpublish.rb: publish public images
This script publishes builds from a previous vmbuild.rb invocation to
any of the configurated public directories. Currently this means Atlas
(for Vagrant images) and GCE (for GCE images).

This code could probably use a similar metadata driven approach as
vmbuild.rb. However for now I believe the current approach of hardcoding
some of the calling conventions and parameters for each of the supported
targets works.