Fix when building from tag
Old code (all run as subshells):

  git reset --hard
  git clean -dxf
  git checkout <ref>
  git fetch origin
  git reset --hard origin/<ref>

The last line only works when the --reference <ref> parameter is a
branch. The build does not fail though, because a subshell failing
doesn't exit the Ruby interpreter. So the code was in essence a NOP in
case of a branch.

However, if I understand the old code correctly, there is another issue.
If you build from a new tag for the first time, the code would actually
not build that tag but whatever was there before. This is because the
"git fetch" is done after the "git checkout <ref>". This could actually
have lead to releases that don't correspond to their tag :-D

The new code looks like this:

  rm -rf /build/references/<ref>
  chdir /build/references/<ref>
  git clone <url> .
  git checkout <ref>

The "rm -rf" should be innocuous. It will never remove your working repo
which is at /build/manageiq-appliance-build. It will only remove a repo
that the script itself checked out. It's safe even if you forget --local.