Add support for building EC2 images.
This builds a VHD image that can be imported using the EC2 VM Import service.

As an alternative to using the VM Import service, the EC2 image can also be
converted to a RAW file, which can then be copied onto an EBS volume and turned
directly into an image. The image will work as-is without further V2V
conversion as the necessary xen network and disk drivers are included in the
initramfs.