Add optional post-build compression step
There are a few current and future appliances where the output of
imagefactory is a non-compressed format. This patch adds support for an
optional post-build compression step to reduce the size of those images.

The supported formats are "gzip" for cases where the users are mostly
Unix/Linux people, and "zip" for Windows folks.

This PR doesn't yet enable any compression. However I would suggest we
enable in a future PR:

* ZIP compression for Azure (given the Windows audience). This will bring
  down the appliance size of 3.3GB to a size similar to the other appliances
  (currently roughtly 1GB for the nightly builds).

* ZIP compression for HyperV. Same situation as Azure.

* GZIP compression for EC2 (once EC2 PR has been enabled). The native image
  format for EC2 is a sparse RAW file which will be 50GB in size. This is
  not practical and must be compressed.