Remove the redis package
We're not in a situation where we can support either a clustered
redis instance or a single instance which all of our appliances
would point at. These cases would require more management and
configuration than we are prepared to tackle.

The distributed case also will not work because the websocket push
notification would be inserted on the redis instance local to the
server which generated it. This would not necessarily be the same
redis instance which the websocket worker would be watching.