Run `bundle clean --force` after we `bundle install`
This will remove any duplicate gems that we have that were installed
with the system ruby (for example, minitest).

This was causing the following warning when starting the appliance_console

WARN: Unresolved specs during Gem::Specification.reset:
      minitest (~> 5.1)
WARN: Clearing out unresolved specs.
Please report a bug if this causes problems.