Don't automatically initialize appliances
For demo purposes it is nice to have an appliance which automatically
initializes itself on boot, but this makes it quite difficult to use
these images in a production deployment which requires multiple
appliances.

This removes the auto-initialization on first-boot allowing the user to
use appliance_console to connect the appliance to an existing region or
create a new one.