Add a script that can be run by cron to clean up factory build vms
If something goes wrong in the VM build process, ImageFactory will give up
after 7200 seconds.  It doesn't clean up that VM, it just takes a screenshot.
If too mnay running VMs are left around it causes problems for new builds
since the host can run out of memory.

CPU time is not perfect, but it should be good enough.